add C extension to support Curve25519
This allows us to drop rbnacl and libsodium, which we've had
problems building on some platforms due to their use of FFI. Also
in this change, I've dropped the "hkdf" gem in favor of just doing
the work ourselves with Ruby's OpenSSL support.

The implementation of Curve25519 comes from Google here:
https://code.google.com/archive/p/curve25519-donna/

This implementation of Curve25519 appears to be in wide use.  It's
also relatively simple - it is less than 1k lines of code.  And it
allows us to drop the much larger libsodium.