Update RSpec setup to use a standard setup
This replaces the RSpec setup with the one generated by `rspec --init`.
The prior gem and webmock requires in the `spec/spec_helper.rb` are
retained to keep the specs passing.

This doesn't really change the setup much. It pulls in RSpec defaults
that are good to have and makes it so that the helper is automatically
required.

A minor update was needed to one spec which was still using the global
`describe` monkey-patch.