Support more versions of Ruby
Test the "latest" of each minor version of Ruby 2.x to ensure
compatibility. Also, explicitly test the exact Ruby version currently
used by the zarkorr project to further ensure there won't be an issue.

:weary: It doesn't appear that Travis-CI has updated their build alias
names (i.e. Ruby 2.2 is only Ruby 2.2.0). This sets up the build matrix
to use the latest release of each minor Ruby 2.x version. To prepare for
the next Ruby release this also includes the `ruby-head` version.
However, as it's a future release we allow it to fail while still
considering the build as passing.