Turn Ruby 2.4+ warnings into build failures
Ruby 2.4 exposed the Ruby warnings method in a new module `Warning`. As
this is a gem we should strive to write warning free code. By using this
new `Warning` hook we can easily convert Ruby warnings into errors thus
failing the build.

To try to simulate this behavior in older Ruby versions we would need to
use an RSpec hook to capture and inspect the error stream output. This
is messy and prone to a variety of problems. As such we are not going to
attempt it and will just have to settle for things reported in 2.4 and
newer.

To play nice we do not overwrite the original Ruby warning behavior.
Instead we use the normal Ruby pattern of creating a custom module and
then extending the Ruby behavior with it. This ensures that if we need
to get access to the Ruby warning behavior at some point we still can.
Additionally, it exposes our monkey patch in the ancestor chain (which
often helps make it easier to debug when things get weird).

As for the error itself, this uses `SyntaxError` instead of the default
`RuntimeError`. I thought the `SyntaxError` was more appropriate for
warnings as they are generally related to the code syntax and structure.
Also, `SyntaxError` inherits from `ScriptError` which is **not** a
`StandardError` subclass. This means that the vast majority of `rescue`
blocks in code won't accidentally intercept the warning error (the
default `rescue` error is `StandardError`).