Fix :bug: with setting cipher key on Ruby 2.4
Ruby 2.4 changed the behavior of `OpenSSL::Cipher#key=` to check the key
size before continuing:

    ArgumentError:
      key must be 16 bytes

In general this is a good thing. When there is a key size mismatch there
could be incorrect underlying assumptions about what the does to the
encryption. For example, since passwords are generally stronger the
longer they are I can easily see someone thinking this applies to the
cipher key as well. But that would depend on the algorithm and is not
generally true for block algorithms. Similarly, if the key is too short
what happens? Is the cipher supposed to pad the key? With what
character? This may also be dependent on the algorithm or may simply not
be allowed.

Prior to Ruby 2.4 it did ensure that the key wasn't shorter than
necessary for the algorithm. This simply compliments that by checking
that it doesn't exceed it either.

When computing the EID value the generation of the temporary key
produces text twice the length of the key. We are expected to truncate
this to 128-bits to use as the key for the final round of computation.
We were relying on the fact that the Ruby AES 128-ECB algorithm
implementation had been doing this for us. However, now that we support
Ruby 2.4 we can no longer rely on that and must perform all truncation
ourselves.

To be robust against future changes this asks the cipher algorithm for
the expected key length and uses that. Ruby will not pad the slice so if
it is not long enough we will receive an error in the future as well.
The spec does not state what should be used to pad the key. In fact
based on the spec we should never hit this situation. If we do, there is
definitely some sort of error and we do want that to be exposed.

See:

- [Eddystone-EID spec](https://github.com/google/eddystone/blob/master/eddystone-eid/eid-computation.md)